# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Mizaan Properties real-estate marketplace and dashboard
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Category = typeof Category[keyof typeof Category];


export const Category = {
  build: 'build',
  apartment: 'apartment',
  land: 'land',
} as const;

export type ListingType = typeof ListingType[keyof typeof ListingType];


export const ListingType = {
  sale: 'sale',
  rent: 'rent',
} as const;

export type RentalPeriod = typeof RentalPeriod[keyof typeof RentalPeriod];


export const RentalPeriod = {
  monthly: 'monthly',
  yearly: 'yearly',
} as const;

export interface Owner {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  additionalContact?: string | null;
}

export interface Location {
  /** @nullable */
  address?: string | null;
  /** @minLength 1 */
  city: string;
  /** @nullable */
  district?: string | null;
}

export interface Listing {
  id: string;
  title: string;
  category: Category;
  listingType: ListingType;
  description: string;
  /** @minimum 0 */
  price: number;
  rentalPeriod: RentalPeriod | null;
  owner: Owner;
  location: Location;
  images: string[];
  createdAt: string;
  updatedAt: string;
}

export interface ListingInput {
  /** @minLength 2 */
  title: string;
  category: Category;
  listingType: ListingType;
  /** @minLength 10 */
  description: string;
  /** @minimum 0 */
  price: number;
  rentalPeriod?: RentalPeriod | null;
  owner: Owner;
  location: Location;
  /** @minItems 1 */
  images: string[];
}

export type ListingUpdate = ListingInput;

export interface ListingSummary {
  total: number;
  builds: number;
  apartments: number;
  lands: number;
  sale: number;
  rent: number;
}

export interface DeleteResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Invalid listing data
 */
export type ValidationErrorResponse = ErrorResponse;

/**
 * Listing not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Server error
 */
export type ServerErrorResponse = ErrorResponse;

export type CategoryParameter = Category;

export type ListingTypeParameter = ListingType;

export type ListListingsParams = {
category?: CategoryParameter;
listingType?: ListingTypeParameter;
/**
 * Search title, city, district, or address
 */
search?: string;
/**
 * Sort field
 */
sort?: ListListingsSort;
/**
 * Sort direction
 */
order?: ListListingsOrder;
};

export type ListListingsSort = typeof ListListingsSort[keyof typeof ListListingsSort];


export const ListListingsSort = {
  createdAt: 'createdAt',
  price: 'price',
} as const;

export type ListListingsOrder = typeof ListListingsOrder[keyof typeof ListListingsOrder];


export const ListListingsOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;