# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the Mizaan Properties real-estate marketplace and dashboard
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: listings
    description: Real-estate listing operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List real-estate listings
      parameters:
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/ListingType"
        - name: search
          in: query
          description: Search title, city, district, or address
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, price]
            default: createdAt
        - name: order
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
  /listings/summary:
    get:
      operationId: getListingSummary
      tags: [listings]
      summary: Get listing summary statistics
      responses:
        "200":
          description: Listing totals grouped by category and listing type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingSummary"
        "500":
          $ref: "#/components/responses/ServerError"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - $ref: "#/components/parameters/ListingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Delete confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  parameters:
    ListingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    Category:
      name: category
      in: query
      schema:
        $ref: "#/components/schemas/Category"
    ListingType:
      name: listingType
      in: query
      schema:
        $ref: "#/components/schemas/ListingType"
  responses:
    ValidationError:
      description: Invalid listing data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Listing not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Category:
      type: string
      enum: [build, apartment, land]
    ListingType:
      type: string
      enum: [sale, rent]
    RentalPeriod:
      type: string
      enum: [monthly, yearly]
    Owner:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        email:
          type: ["string", "null"]
        additionalContact:
          type: ["string", "null"]
      required: [name, phone]
    Location:
      type: object
      properties:
        address:
          type: ["string", "null"]
        city:
          type: string
          minLength: 1
        district:
          type: ["string", "null"]
      required: [city]
    Listing:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        listingType:
          $ref: "#/components/schemas/ListingType"
        description:
          type: string
        price:
          type: number
          minimum: 0
        rentalPeriod:
          oneOf:
            - $ref: "#/components/schemas/RentalPeriod"
            - type: "null"
        owner:
          $ref: "#/components/schemas/Owner"
        location:
          $ref: "#/components/schemas/Location"
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - listingType
        - description
        - price
        - rentalPeriod
        - owner
        - location
        - images
        - createdAt
        - updatedAt
    ListingInput:
      type: object
      properties:
        title:
          type: string
          minLength: 2
        category:
          $ref: "#/components/schemas/Category"
        listingType:
          $ref: "#/components/schemas/ListingType"
        description:
          type: string
          minLength: 10
        price:
          type: number
          minimum: 0
        rentalPeriod:
          oneOf:
            - $ref: "#/components/schemas/RentalPeriod"
            - type: "null"
        owner:
          $ref: "#/components/schemas/Owner"
        location:
          $ref: "#/components/schemas/Location"
        images:
          type: array
          minItems: 1
          items:
            type: string
      required:
        - title
        - category
        - listingType
        - description
        - price
        - owner
        - location
        - images
    ListingUpdate:
      allOf:
        - $ref: "#/components/schemas/ListingInput"
    ListingSummary:
      type: object
      properties:
        total:
          type: number
        builds:
          type: number
        apartments:
          type: number
        lands:
          type: number
        sale:
          type: number
        rent:
          type: number
      required: [total, builds, apartments, lands, sale, rent]
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]