# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Mizaan Properties real-estate marketplace and dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { ListingType } from './listingType';
import type { Location } from './location';
import type { Owner } from './owner';
import type { RentalPeriod } from './rentalPeriod';

export interface ListingInput {
  /** @minLength 2 */
  title: string;
  category: Category;
  listingType: ListingType;
  /** @minLength 10 */
  description: string;
  /** @minimum 0 */
  price: number;
  rentalPeriod?: RentalPeriod | null;
  owner: Owner;
  location: Location;
  /** @minItems 1 */
  images: string[];
}