# lib/api-zod/src/generated/types/listingSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Mizaan Properties real-estate marketplace and dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface ListingSummary {
  total: number;
  builds: number;
  apartments: number;
  lands: number;
  sale: number;
  rent: number;
}