# lib/api-zod/src/generated/types/listingType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Mizaan Properties real-estate marketplace and dashboard
 * OpenAPI spec version: 0.1.0
 */

export type ListingType = typeof ListingType[keyof typeof ListingType];


export const ListingType = {
  sale: 'sale',
  rent: 'rent',
} as const;