# lib/api-zod/src/generated/types/listingUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Mizaan Properties real-estate marketplace and dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { ListingInput } from './listingInput';

export type ListingUpdate = ListingInput;