# lib/api-zod/src/generated/types/location.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Mizaan Properties real-estate marketplace and dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface Location {
  /** @nullable */
  address?: string | null;
  /** @minLength 1 */
  city: string;
  /** @nullable */
  district?: string | null;
}