# lib/api-zod/src/generated/types/owner.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Mizaan Properties real-estate marketplace and dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface Owner {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  additionalContact?: string | null;
}