# lib/api-zod/src/generated/types/rentalPeriod.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Mizaan Properties real-estate marketplace and dashboard
 * OpenAPI spec version: 0.1.0
 */

export type RentalPeriod = typeof RentalPeriod[keyof typeof RentalPeriod];


export const RentalPeriod = {
  monthly: 'monthly',
  yearly: 'yearly',
} as const;